vmamba-lab/requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Mamba
mamba-ssm>=1.2.0
causal-conv1d>=1.2.0

# Wavelet Transform
PyWavelets>=1.4.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Model utilities
timm>=0.9.0
einops>=0.7.0

# Face Detection
facenet-pytorch>=2.5.0
# insightface>=0.7.0  # Optional, for better face detection

# Logging & Experiment Tracking
wandb>=0.15.0
tensorboard>=2.14.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Evaluation
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0

# Deformable Convolution (for DCConv)
# torchvision already includes deform_conv2d

# Optional: ONNX export
# onnx>=1.14.0
# onnxruntime>=1.15.0
